requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
requests
jinja2
fastapi-mail
alembic
email-validator
python-dotenv
bcrypt
jose
scikit-image
torch
opencv-python
imagehash
torchvision